basic output

+++ lib/active_record/annotate/writer.rb
@@ -3,20 +3,27 @@ module ActiveRecord
     module Writer
       class << self
         def write(annotation, path)
-          new_file_content = assemble(annotation, path)
+          original_file_content = File.read(path)
+          lines = original_file_content.split("\n")
           
-          temp_path = "#{path}.annotated"
-          File.open(temp_path, 'w') do |temp_file|
-            temp_file.puts(new_file_content)
+          processed_lines = process(lines, annotation)
+          new_file_content = processed_lines.join("\n")
+          
+          if new_file_content != original_file_content
+            temp_path = "#{path}.annotated"
+            File.open(temp_path, 'w') do |temp_file|
+              temp_file.write(new_file_content)
+            end
+            
+            File.delete(path)
+            File.rename(temp_path, path)
+            
+            relative_path = path.sub(/^#{Rails.root}\//, '')
+            puts "* #{relative_path}"
           end
-          
-          File.delete(path)
-          File.rename(temp_path, path)
         end
         
-        def assemble(annotation, path)
-          lines = File.read(path).split("\n")
-          
+        def process(lines, annotation)
           if lines.first =~ /^\s*#.*coding/
             # encoding: utf-8 encountered on the first line
             encoding_line = lines.shift
@@ -27,10 +34,11 @@ module ActiveRecord
             lines.shift
           end
           
-          lines.unshift(*annotation, nil)
-          lines.unshift(encoding_line) unless encoding_line.nil?
-          
-          lines.join("\n")
+          lines.tap do |lines|
+            lines.unshift(*annotation, nil)
+            lines.unshift(encoding_line) unless encoding_line.nil?
+            lines.push(nil) # newline at the end of file
+          end
         end
       end
     end
